refactor(desktop): split large workspace hooks
