+++ apps/desktop/src/hooks/workspaces/workflows/tabs/workspace-shell-activation-types.ts
@@ -1,0 +1,9 @@
+import type { MeasurementOperationId } from "@/lib/domain/telemetry/debug-measurement-catalog";
+
+export interface SelectSessionOptionsWithoutGuard {
+  latencyFlowId?: string | null;
+  measurementOperationId?: MeasurementOperationId | null;
+  reuseMeasurementOperation?: boolean;
+  allowColdIdleNoStream?: boolean;
+  forceCold?: boolean;
+}
